+++ Sources/LinkPreview/LinkPreviewFailure.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+enum LinkPreviewFailure: Error {
+    case noURLFound
+    case invalidURL
+    case cannotBeOpened
+    case parseError
+}