+++ CHANGELOG.md
@@ -10,6 +10,39 @@ The changelog for [ApplozicSwift](https://github.com/AppLozic/ApplozicSwift). Al
 ### Enhancements
 
 - [CM-193] Added support to open the application when tapping any URL in the text message. If the application isn't installed, we'll open it in Safari.
+- Show link preview in chat for web links
+</br>You can use below config to disable web link preview in chat:
+```
+    config.isLinkPreviewDisabled = true
+```
+You need to set `Allow Arbitrary` Loads to YES on your project's `Info.plist` file.
+```
+ <key>NSAppTransportSecurity</key>
+  <dict>
+      <key>NSAllowsArbitraryLoads</key>
+      <true/>
+  </dict>
+```
+If you don't want to use the above option, You can white list some the websites which you want to allow them in `Info.plist` file
+```
+<key>NSAppTransportSecurity</key>
+<dict>
+  <key>NSAllowsArbitraryLoads</key>
+  <false/>
+  <key>NSExceptionDomains</key>
+  <dict>
+    <key>google.com</key>
+    <dict>
+      <key>NSIncludesSubdomains</key>
+      <true/>
+      <key>NSExceptionAllowsInsecureHTTPLoads</key>
+      <true/>
+    </dict>
+  </dict>
+</dict>
+```
+You can read more on Property List Keys
+[here](https://developer.apple.com/documentation/bundleresources/information_property_list/nsapptransportsecurity/nsexceptiondomains)
 
 ### Fixes